Changes dependencies versions.
Should fix #10.

Changes html-pipeline version to latest (0.0.6) and linguist version to 2.1
which is the one html-pipeline is using.

If you get an error similar to this:
`block in ffi_lib': Could not open library 'lib.so'`
It's probably related to the Python version. It was failing for me when using
Python 3, but started working when I switched to Python 2.